feat: Move states to contexts
